requirements.txt
Django=2.2.7
Pillow==5.4.1
djangorestframework==3.11.0
moviepy==1.0.3
gunicorn==20.0.4
django-heroku
